backend/requirements.txt
Django>=4
django-extensions
djangorestframework
django-cors-headers